Move into Domain.Arena namespace
