+++ WalletWasabi/Domains/Arena/Events/InputsConnectionConfirmationStartedEvent.cs
@@ -5,10 +5,10 @@ using System.Text;
 using System.Threading.Tasks;
 using NBitcoin;
 using WalletWasabi.Crypto;
+using WalletWasabi.Domains.Arena.Interfaces;
 using WalletWasabi.EventSourcing.Interfaces;
-using WalletWasabi.WabiSabi.Backend.Rounds;
 
-namespace WalletWasabi.EventSourcing.ArenaDomain.Events
+namespace WalletWasabi.Domains.Arena.Events
 {
-	public record InputRegisteredEvent(Guid AliceSecret, Coin Coin, OwnershipProof OwnershipProof) : IEvent;
+	public record InputsConnectionConfirmationStartedEvent() : IEvent, IRoundClientEvent;
 }
