+++ WalletWasabi/Domains/Arena/Events/OutputRegisteredEvent.cs
@@ -5,9 +5,10 @@ using System.Text;
 using System.Threading.Tasks;
 using NBitcoin;
 using WalletWasabi.Crypto;
+using WalletWasabi.Domains.Arena.Interfaces;
 using WalletWasabi.EventSourcing.Interfaces;
 
-namespace WalletWasabi.EventSourcing.ArenaDomain.Events
+namespace WalletWasabi.Domains.Arena.Events
 {
-	public record InputConnectionConfirmedEvent(Coin Coin, OwnershipProof OwnershipProof) : IEvent, IRoundClientEvent;
+	public record OutputRegisteredEvent(Script Script, long CredentialAmount) : IEvent, IRoundClientEvent;
 }
