+++ WalletWasabi/Domains/Arena/Events/RoundEndedEvent.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using NBitcoin;
+using WalletWasabi.Domains.Arena.Interfaces;
 using WalletWasabi.EventSourcing.Interfaces;
 
-namespace WalletWasabi.EventSourcing.ArenaDomain.Events
+namespace WalletWasabi.Domains.Arena.Events
 {
-	public record RoundSucceedEvent() : IEvent, IRoundClientEvent;
+	public record RoundEndedEvent() : IEvent, IRoundClientEvent;
 }
