+++ WalletWasabi/Domains/Arena/Events/RoundStartedEvent.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using WalletWasabi.EventSourcing.ArenaDomain.Aggregates;
+using WalletWasabi.Domains.Arena.Aggregates;
+using WalletWasabi.Domains.Arena.Interfaces;
 using WalletWasabi.EventSourcing.Interfaces;
 using WalletWasabi.WabiSabi.Backend.Rounds;
 
-namespace WalletWasabi.EventSourcing.ArenaDomain.Events
+namespace WalletWasabi.Domains.Arena.Events
 {
 	public record RoundStartedEvent(RoundParameters2 RoundParameters) : IEvent, IRoundClientEvent;
 }
