+++ WalletWasabi/Domains/Arena/Events/SignatureAddedEvent.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using NBitcoin;
+using WalletWasabi.Domains.Arena.Interfaces;
+using WalletWasabi.EventSourcing.Interfaces;
+
+namespace WalletWasabi.Domains.Arena.Events
+{
+	public record SignatureAddedEvent(OutPoint AliceOutPoint, WitScript WitScript) : IEvent, IRoundClientEvent;
+}
