+++ WalletWasabi/Domains/Arena/Events/SigningStartedEvent.cs
@@ -3,9 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using WalletWasabi.Domains.Arena.Interfaces;
 using WalletWasabi.EventSourcing.Interfaces;
 
-namespace WalletWasabi.EventSourcing.ArenaDomain.Events
+namespace WalletWasabi.Domains.Arena.Events
 {
-	public record OutputRegistrationStartedEvent() : IEvent, IRoundClientEvent;
+	public record SigningStartedEvent() : IEvent, IRoundClientEvent;
 }
