+++ WalletWasabi/Domains/Arena/Interfaces/IRoundClientEvent.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WalletWasabi.EventSourcing.ArenaDomain.Events
+namespace WalletWasabi.Domains.Arena.Interfaces
 {
 	public interface IRoundClientEvent
 	{
