--- WalletWasabi/EventSourcing/ArenaDomain/Events/RoundEndedEvent.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using WalletWasabi.EventSourcing.Interfaces;
-
-namespace WalletWasabi.EventSourcing.ArenaDomain.Events
-{
-	public record RoundEndedEvent() : IEvent, IRoundClientEvent;
-}
